refactor: consolidate scripts, add hashmap benchmarks & quick reference
- Relocate all utility scripts to scripts/ directory
- Add hashmap_bench.cc with performance benchmarking suite
- Introduce PERFORMANCE_QUICKREF.md for at-a-glance metrics
- Update CI workflows and documentation to reflect new structure

+++ src/twopointers.cc
@@ -1515,7 +1515,7 @@ int Purgatory::maxNumOfMarkedIndices(vector<int> &nums) {
   int pair = 0;
 
   while (left < limit && right < n) {
-    const long long twice = nums[left] << 1;
+    const long long twice = static_cast<long long>(nums[left]) << 1;
     const int large = nums[right];
 
     // branch prediction
